@@ -1,64 +1,54 @@
 {
-  "name": "@restqa/editor",
-  "version": "4.0.0",
+  "name": "dashboard",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build",
-    "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint",
     "test": "vue-cli-service test:unit",
     "test:watch": "vue-cli-service test:unit --watch"
   },
   "dependencies": {
-    "@chenfengyuan/vue-countdown": "^1.1.2",
     "axios": "^0.21.1",
-    "core-js": "2.6.5",
-    "material-icons": "^0.3.1",
-    "perfect-scrollbar": "^1.4.0",
-    "postcss-rtl": "^1.5.0",
-    "prismjs": "^1.16.0",
-    "vee-validate": "^2.2.8",
-    "vue": "^2.6.10",
-    "vue-awesome-swiper": "^3.1.3",
-    "vue-backtotop": "^1.6.1",
-    "vue-clipboard2": "^0.3.0",
-    "vue-context": "4.0.0",
-    "vue-feather-icons": "^5.0.0",
-    "vue-flatpickr-component": "^8.1.2",
-    "vue-i18n": "^8.11.2",
-    "vue-perfect-scrollbar": "^0.1.0",
-    "vue-prism-component": "^1.1.1",
-    "vue-property-decorator": "^8.1.1",
-    "vue-quill-editor": "^3.0.6",
-    "vue-router": "^3.0.6",
-    "vue-select": "^3.1.0",
-    "vue-simple-suggest": "^1.9.5",
-    "vue-socket.io": "^3.0.10",
-    "vue-tree-halower": "^1.8.0",
-    "vue2-hammer": "^2.1.2",
-    "vuedraggable": "^2.21.0",
-    "vuesax": "^3.12.2",
-    "vuex": "^3.1.1"
+    "core-js": "^3.6.5",
+    "element-plus": "^1.0.2-beta.45",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.8",
+    "vuex": "^4.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.0",
-    "@babel/preset-env": "^7.14.1",
-    "@fullhuman/postcss-purgecss": "^1.2.0",
-    "@vue/cli-plugin-babel": "^3.7.0",
-    "@vue/cli-plugin-eslint": "^3.7.0",
-    "@vue/cli-plugin-unit-jest": "^4.5.12",
-    "@vue/cli-service": "^3.7.0",
-    "@vue/test-utils": "^1.0.3",
-    "axios-mock-adapter": "^1.17.0",
-    "babel-core": "^7.0.0-bridge.0",
-    "lodash": "^4.17.21",
-    "node-sass": "^4.12.0",
-    "purgecss": "^1.3.0",
-    "sass-loader": "^7.1.0",
-    "script-loader": "0.7.2",
-    "tailwindcss": "^1.0.1",
-    "vue-jest": "^3.0.7",
-    "vue-template-compiler": "^2.6.10"
-  }
+    "@chenfengyuan/vue-countdown": "^2.0.0-beta",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-unit-jest": "^4.5.13",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.11",
+    "@vue/test-utils": "^2.0.0-rc.6",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^6.7.2",
+    "eslint-plugin-vue": "^7.0.0",
+    "sass": "^1.34.0",
+    "sass-loader": "^10.2.0",
+    "vue-jest": "^5.0.0-alpha.10"
+  },
+  "eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/vue3-essential",
+      "eslint:recommended"
+    ],
+    "parserOptions": {
+      "parser": "babel-eslint"
+    },
+    "rules": {}
+  },
+  "browserslist": [
+    "> 1%",
+    "last 2 versions",
+    "not dead"
+  ]
 }
