@@ -1,64 +1,58 @@
 <template>
-  <vx-card title="Project Initialiization" emoji="🚀"  class="mt-base" slot="no-body" >
-    <div class="alert" :class="alert.type" v-if="alert.text">{{ alert.text }}</div>
-    <label for="name">Name: </label>
-    <input type="text" id="name" v-model="model.name" />
-    <br />
-    <label for="description">Description: </label>
-    <input type="text" id="description" v-model="model.description" />
-    <br />
-    <label for="url">Url: </label>
-    <input type="text" id="url" v-model="model.url" />
-    <br />
-    <label for="env">Environment: </label>
-    <input type="text" id="env" v-model="model.env" />
-    <br />
-    <button id="btn" @click="submit">Create</button>
-    <Loader :show="loader" />
-  </vx-card>
-
+  <card title="Project Initialiization" emoji="🚀" :loading="loader">
+    <el-form ref="form" :model="form" label-width="120px">
+      <el-form-item label="Name">
+        <el-input v-model="form.name"></el-input>
+      </el-form-item>
+      <el-form-item label="Description">
+        <el-input v-model="form.description"></el-input>
+      </el-form-item>
+      <el-form-item label="API Url">
+        <el-input v-model="form.url"></el-input>
+      </el-form-item>
+      <el-form-item label="Environment">
+        <el-input v-model="form.env"></el-input>
+      </el-form-item>
+      <el-form-item>
+        <el-button type="primary" @click="onSubmit">Create</el-button>
+      </el-form-item>
+    </el-form>
+  </card>
 </template>
 
 <script>
-import VxCard from '../../global/vx-card/VxCard'
-import Loader from '../../utils/loader/Loader'
+import Card from '@/components/UI/card/Card'
 import * as Service from '@/services/restqa/project'
 import { ValidationError } from '../../../services/http'
 
 export default {
   name: 'RestQAProjectInit',
   components: {
-    Loader,
-    VxCard
-  },
-  props: {
+    Card
   },
   data () {
     return {
       loader: false,
-      model: {},
-      alert: {
-        text: null,
-        type: null
-      },
-      Service
+      form: {},
     }
   },
   methods: {
-    async submit () {
+    async onSubmit () {
       this.loader = true
       try {
-        await this.Service.initialize(this.model)
-        this.alert = {
-          text: '🚀🚀 Your project has been created successfully!',
+        await Service.initialize(this.form)
+        this.$notify({
+          title: 'Let\'s go!',
+          message: '🚀🚀 Your project has been created successfully!',
           type: 'success'
-        }
+        })
         this.$store.dispatch('config')
       } catch (e) {
-        this.alert = {
-          text: e.message,
+        this.$notify({
+          title: 'Oups',
+          message: e.message,
           type: (e instanceof ValidationError) ? 'warning' : 'error'
-        }
+        })
       }
       this.loader = false
     }
