@@ -1,30 +1,29 @@
-import { shallowMount } from '@vue/test-utils'
+import { mount } from '@vue/test-utils'
+import { createStore } from 'vuex'
+import ElementPlus from 'element-plus';
 import RestQAProjectSteps from './RestQAProjectSteps.vue'
-import Loader from '../../utils/loader/Loader'
-import VxCard from '../../global/vx-card/VxCard'
-
-afterEach(() => {
-  jest.resetModules()
-  jest.resetAllMocks()
-})
 
 describe('RestQAProjectSteps', () => {
+
   test('renders the step definition, when the props data is not defined', async () => {
 
-    const component = shallowMount(RestQAProjectSteps, {
-      propsData: {
+    const options = {
+      props: {
         keyword: 'given'
-      }
-    })
+      },
+      global: {
+        plugins: [
+          ElementPlus
+        ]
+      },
+    }
 
+    const component = mount(RestQAProjectSteps, options)
     expect(component.exists()).toBeTruthy()
 
-    expect(component.findComponent(Loader).isVisible()).toBeTruthy()
-    expect(component.findComponent(VxCard).attributes('title')).toBe('Keyword: Given')
-    expect(component.find('.step-definitions').exists()).toBeFalsy()
+    expect(component.findComponent({ name: 'card'}).vm.title).toBe('Given')
+    expect(component.findComponent({ name: 'card'}).vm.loading).toBe(true)
     
-    await component.vm.$nextTick()
-
     const data = [{
       plugin: '@restqa/restqapi',
       keyword: 'Given',
@@ -41,18 +40,14 @@ describe('RestQAProjectSteps', () => {
 
     await component.vm.$nextTick()
 
-    expect(component.findComponent(Loader).props('show')).toBeFalsy()
-    expect(component.find('.step-definitions').isVisible()).toBeTruthy()
+    expect(component.findComponent({ name: 'card'}).vm.title).toBe('Given - 2 step definitions')
+    expect(component.findComponent({ name: 'card'}).vm.loading).toBe(false)
 
-    const expectedSteps = component.findAll('.step-definition')
-    expect(expectedSteps.at(0).find('.plugin').text()).toEqual('@restqa/restqapi')
-    expect(expectedSteps.at(0).find('.keyword').text()).toEqual('Given')
-    expect(expectedSteps.at(0).find('.step').text()).toEqual('I have the api gateway')
-    expect(expectedSteps.at(0).find('.comment').text()).toEqual('Initiate the api call')
-    expect(expectedSteps.at(1).find('.plugin').text()).toEqual('@restqa/faker-plugin')
-    expect(expectedSteps.at(1).find('.keyword').text()).toEqual('Given')
-    expect(expectedSteps.at(1).find('.step').text()).toEqual('I have the cookie')
-    expect(expectedSteps.at(1).find('.comment').text()).toEqual('Adding the cookie')
+    await component.vm.$nextTick()
+
+    const expectedSteps = component.findComponent('.step-definition')
+    expect(expectedSteps.findComponent('.step').vm.prop).toEqual('step')
+    expect(expectedSteps.findComponent('.step').vm.label).toEqual('Step Definition')
   })
 
   test('renders the step definition, when the props data is passed', async () => {
@@ -69,29 +64,29 @@ describe('RestQAProjectSteps', () => {
       comment: 'Adding the cookie'
     }]
 
-    const component = shallowMount(RestQAProjectSteps, {
-      propsData: {
+    const options = {
+      props: {
         keyword: 'given',
         data
+      },
+      global: {
+        plugins: [
+          ElementPlus
+        ]
       }
-    })
+    }
 
+    const component = mount(RestQAProjectSteps, options)
     expect(component.exists()).toBeTruthy()
-
     await component.vm.$nextTick()
 
-    expect(component.findComponent(Loader).props('show')).toBeFalsy()
-    expect(component.find('.step-definitions').isVisible()).toBeTruthy()
+    expect(component.findComponent({ name: 'card'}).vm.title).toBe('Given - 2 step definitions')
+    expect(component.findComponent({ name: 'card'}).vm.loading).toBe(false)
 
-    const expectedSteps = component.findAll('.step-definition')
-    expect(expectedSteps.at(0).find('.plugin').text()).toEqual('@restqa/restqapi')
-    expect(expectedSteps.at(0).find('.keyword').text()).toEqual('Given')
-    expect(expectedSteps.at(0).find('.step').text()).toEqual('I have the api gateway')
-    expect(expectedSteps.at(0).find('.comment').text()).toEqual('Initiate the api call')
-    expect(expectedSteps.at(1).find('.plugin').text()).toEqual('@restqa/faker-plugin')
-    expect(expectedSteps.at(1).find('.keyword').text()).toEqual('Given')
-    expect(expectedSteps.at(1).find('.step').text()).toEqual('I have the cookie')
-    expect(expectedSteps.at(1).find('.comment').text()).toEqual('Adding the cookie')
+    const expectedSteps = component.findComponent('.step-definition')
+
+    expect(expectedSteps.findComponent('.step').vm.prop).toEqual('step')
+    expect(expectedSteps.findComponent('.step').vm.label).toEqual('Step Definition')
   })
 })
 
