@@ -1,29 +1,33 @@
 <template>
-  <vx-card :title="title" class="mt-base" slot="no-body" >
-    <Loader :show="!data.length" />
-    <div class="step-definitions" v-if="data.length">
-      <div class="step-definition" :key="index" v-for="(step, index) in data">
-        <div class="plugin">{{ step.plugin }}</div>
-        <div class="keyword">{{ step.keyword }}</div>
-        <prism class="step" language="gherkin">{{ step.step }}</prism>
-        <div class="comment">{{ step.comment }}</div>
-      </div>
-    </div>
-  </vx-card>
+  <card :title="title" :loading="null === data">
+    <el-table :data="data" class="step-definition" style="width: 100%">
+      <el-table-column class="expand" type="expand">
+        <template class="ee" #default="props">
+          <el-descriptions title="Detaild"  :column="1" border>
+            <el-descriptions-item label="Keyword">{{ props.row.keyword }}</el-descriptions-item>
+            <el-descriptions-item label="Comment">{{ props.row.comment }}</el-descriptions-item>
+            <el-descriptions-item label="Plugin" class="plugin">{{ props.row.plugin }}</el-descriptions-item>
+          </el-descriptions>
+       </template>
+      </el-table-column>
+      <el-table-column class="step" prop="step" label="Step Definition"></el-table-column>
+      <el-table-column fixed="right" label="Operations" width="120">
+        <template #default>
+          <el-button type="text" size="small">Copy</el-button>
+      </template>
+    </el-table-column>
+    </el-table>
+  </card>
+  <br />
 </template>
 
 <script>
-import Prism from 'vue-prism-component'
-import VxCard from '../../global/vx-card/VxCard'
-import Loader from '../../utils/loader/Loader'
-import * as Service from '@/services/restqa/project'
+import Card from '@/components/UI/card/Card'
 
 export default {
   name: 'RestQAProjectSteps',
   components: {
-    VxCard,
-    Loader,
-    Prism
+    Card
   },
   props: {
     keyword: {
@@ -31,19 +35,13 @@ export default {
     },
     data: {
       type: Array,
-      default: () => []
-    }
-  },
-  data () {
-    return {
-      steps: [],
-      searchTerm: '',
-      Service
+      default: null
     }
   },
   computed: {
     title () {
-      return `Keyword: ${ this.keyword[0].toUpperCase() + this.keyword.substring(1) }`
+       const suffix = (this.data && ` - ${this.data.length} step definitions`)|| ''
+      return `${ this.keyword[0].toUpperCase() + this.keyword.substring(1)}${suffix}`
     }
   }
 }
