@@ -1,89 +1,59 @@
-import Vue from 'vue'
-import Router from 'vue-router'
+import { createRouter, createWebHashHistory } from 'vue-router'
+import Store from './store/store'
 
-Vue.use(Router)
-
-const router = new Router({
-  mode: 'hash',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHashHistory(process.env.BASE_URL),
   scrollBehavior () {
-    return { x: 0, y: 0 }
-  },
-  routes: [
-
-    {
-    // =============================================================================
-    // MAIN LAYOUT ROUTES
-    // =============================================================================
-      path: '',
-      component: () => import('./layouts/main/Main.vue'),
-      children: [
-        // =============================================================================
-        // Theme Routes
-        // =============================================================================
-        {
-          path: '/',
-          redirect: '/comingsoon'
-        },
-        {
-          //path: '/',
-          path: '/dashboard',
-          name: 'dashboard-restgqa',
-          component: () => import('./views/DashboardRestQA.vue'),
-          meta: {
-            rule: 'editor'
-          }
-        },
-        {
-          path: '/project/steps',
-          name: 'project-steps',
-          component: () => import('./views/restqa/steps/RestQASteps.vue')
-        }
-      ]
-    },
-    // =============================================================================
-    // FULL PAGE LAYOUTS
-    // =============================================================================
-    {
-      path: '',
-      component: () => import('@/layouts/full-page/FullPage.vue'),
-      children: [
-        {
-          path: '/comingsoon',
-          name: 'page-coming-soon',
-          component: () => import('@/views/pages/ComingSoon.vue'),
-          meta: {
-            rule: 'editor'
-          }
-        },
-        {
-          path: '/error-404',
-          name: 'page-error-404',
-          component: () => import('@/views/pages/Error404.vue'),
-          meta: {
-            rule: 'editor'
-          }
-        }
-      ]
-    },
-    // Redirect to 404 page, if no match found
-    {
-      path: '*',
-      redirect: '/pages/error-404'
+    return {
+      left: 0,
+      top: 0
     }
-  ]
+  },
+  routes: [{
+    path: '/',
+    component: () => import('./layouts/FullPage.vue'),
+    children: [{
+      path: '',
+      name: 'page-coming-soon',
+      component: () => import('./views/pages/ComingSoon.vue'),
+      meta: {
+        title: 'Coming soon!'
+      }
+    }]
+  }, {
+    path: '/beta',
+    component: () => import('./layouts/Dashboard.vue'),
+    children: [{
+        path: '',
+        name: 'homepage',
+        component: () => import('./views/Homepage.vue'),
+        meta: {
+          title: 'Dashboard'
+        }
+      }, {
+        path: '/steps',
+        name: 'steps',
+        component: () => import('./views/Steps/Steps.vue'),
+        meta: {
+          title: 'Step definition'
+        }
+      }
+    ]
+  }]
 })
 
-router.afterEach(() => {
-  // Remove initial loading
+router.beforeEach((to) => {
+  if (to.name !== 'homepage' && null === Store.getters.config) {
+    //return {'name': 'homepage'}
+  }
+})
+
+router.afterEach((to) => {
+  document.title = (to.meta.title || 'Testing with ❤️') + ' | RestQA Dashboard 🦏'
   const appLoading = document.getElementById('loading-bg')
   if (appLoading) {
     appLoading.style.display = 'none'
   }
-})
-
-router.beforeEach((to, from, next) => {
-  next()
 })
 
 export default router
