@@ -1,7 +1,9 @@
 const state = {
   config: null,
   loadingConfig: false,
-  selectedEnv: null
+  selectedEnv: null,
+  info: null,
+  steps: null
 }
 
 export default state
