@@ -1,20 +1,7 @@
-import Vue from 'vue'
-import Vuex from 'vuex'
-
-import state from './state'
-import getters from './getters'
-import mutations from './mutations'
-import actions from './actions'
+import { createStore } from 'vuex'
 import restqa from './modules/restqa'
 
-Vue.use(Vuex)
-
-
-export default new Vuex.Store({
-  getters,
-  mutations,
-  state,
-  actions,
+export default createStore({
   modules: {
     restqa
   },
